Refactor settings structure and enhance authentication utilities
- Moved settings import from app.settings to app.core.settings across multiple files.
- Introduced new auth.py module for API documentation authentication utilities.
- Added settings for documentation authentication, including username, password, and enable flag.
- Improved error handling in database session management.
